addressing PR feedback: refactor subinterfaces/implementations into their own files, revert accidental settings.gradle.kts change, clean up various KDocs
